Improve filterset application using custom backend
Fixes #27
